pyproject.toml
[project]
name = "mrp-james"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "altgraph>=0.17,<0.18",
    "asgiref>=3.8,<3.9",
    "blinker>=1.9,<1.10",
    "bottle>=0.13,<0.14",
    "django>=4.2,<4.3",
    "django-appconf>=1.1,<1.2",
    "django-axes>=8.3.1",
    "django-import-export>=4.3,<4.4",
    "djangorestframework>=3.16,<3.17",
    "flask>=3.1,<3.2",
    "getmac>=0.9,<0.10",
    "idna>=3.10,<3.11",
    "importlib-resources>=6.5,<6.6",
    "jinja2>=3.1,<3.2",
    "json5>=0.12,<0.13",
    "markupsafe>=3.0,<3.1",
    "numpy>=2.3,<2.4",
    "odfpy>=1.4,<1.5",
    "openpyxl>=3.1,<3.2",
    "packaging>=25.0,<25.1",
    "pandas>=2.3,<2.4",
    "pathspec>=0.12,<0.13",
    "pefile>=2023.2,<2023.3",
    "pillow>=10.4,<10.5",
    "pycodestyle>=2.14,<2.15",
    "pyflakes>=3.4,<3.5",
    "pypdf2>=3.0,<3.1",
    "python-dateutil==2.9.0.post0",
    "python-dotenv>=1.0,<1.1",
    "python-json-logger>=3.3,<3.4",
    "pytz>=2025.2,<2025.3",
    "pywin32>=311,<312 ; sys_platform == 'win32'",
    "pywin32-ctypes>=0.2,<0.3",
    "pyyaml>=6.0,<6.1",
    "regex>=2024.11,<2024.12",
    "requests>=2.32,<2.33",
    "setuptools>=80.9,<80.10",
    "six>=1.17,<1.18",
    "sqlparse>=0.5,<0.6",
    "tablib>=3.8,<3.9",
    "tqdm>=4.67,<4.68",
    "typing-extensions>=4.14,<4.15",
    "tzdata>=2025.2,<2025.3",
    "urllib3>=2.5,<2.6",
    "werkzeug>=3.1,<3.2",
    "xlrd>=2.0,<2.1",
    "xlwt>=1.3,<1.4",
    "zope-event>=5.1,<5.2",
    "zope-interface>=7.2,<7.3",
]
